+++ index.js
@@ -1,4 +1,1 @@
-import { usePaging } from "./src/ss-paging-esm";
-import { usePagingStore } from "./src/ss-paging-store";
-
-export { usePaging, usePagingStore }
+export { usePaging, usePagingStore } from "./src/ss-paging-esm";
